Extract control flow package. Refactor codes.
