+++ Source/Commands/Help.php
@@ -4,7 +4,7 @@ namespace Phpkg\Commands\Help;
 
 use PhpRepos\Cli\IO\Write;
 
-function run(): void
+return function (): void
 {
     $content = <<<EOD
 usage: phpkg [-v | --version] [-h | --help] [--man]
@@ -36,4 +36,4 @@ global access
 EOD;
 
     Write\line($content);
-}
+};
