+++ Source/Commands/Init.php
@@ -2,6 +2,7 @@
 
 namespace Phpkg\Commands\Init;
 
+use Phpkg\Application\PackageManager;
 use Phpkg\Classes\Config\Config;
 use Phpkg\Classes\Meta\Meta;
 use Phpkg\Classes\Environment\Environment;
@@ -10,12 +11,12 @@ use Phpkg\Exception\PreRequirementsFailedException;
 use PhpRepos\FileManager\Directory;
 use PhpRepos\FileManager\File;
 use PhpRepos\FileManager\Filename;
-use PhpRepos\FileManager\JsonFile;
 use function PhpRepos\Cli\IO\Read\parameter;
 use function PhpRepos\Cli\IO\Write\line;
 use function PhpRepos\Cli\IO\Write\success;
+use function PhpRepos\ControlFlow\Transformation\pipe;
 
-function run(Environment $environment): void
+return function (Environment $environment): void
 {
     line('Init project...');
     $project = new Project($environment->pwd->append(parameter('project', '')));
@@ -30,15 +31,12 @@ function run(Environment $environment): void
         return $config;
     });
 
-    $meta = Meta::init();
-
     $project->config($config);
-    $project->meta = $meta;
+    $project->meta = Meta::init();;
 
-    JsonFile\write($project->config_file, $project->config->to_array());
-    JsonFile\write($project->meta_file, $project->meta->to_array());
+    PackageManager\commit($project);
 
     Directory\exists_or_create($project->packages_directory);
 
     success('Project has been initialized.');
-}
+};
