+++ Source/Commands/Version.php
@@ -4,7 +4,7 @@ namespace Phpkg\Commands\Version;
 
 use PhpRepos\Cli\IO\Write;
 
-function run(): void
+return function (): void
 {
     Write\success('phpkg version 1.2.0');
-}
+};
