+++ Tests/System/BuildCommand/BuildCommandTest.php
@@ -71,16 +71,7 @@ function assert_output($output)
 \e[39mStart building...
 \e[39mLoading configs...
 \e[39mChecking packages...
-\e[39mPrepare build directory...
-\e[39mMake namespaces map...
-\e[39mBuilding packages...
-\e[39mBuilding package git@github.com:php-repos/simple-package.git...
-\e[39mBuilding the project...
-\e[39mBuilding entry points...
-\e[39mBuilding entry point entry-point
-\e[39mBuilding executables...
-\e[39mBuilding executables for package git@github.com:php-repos/simple-package.git
-\e[39mBuilding executable file run-executable from run.php
+\e[39mBuilding...
 \e[92mBuild finished successfully.\e[39m
 
 EOD;
