+++ phpkg
@@ -2,34 +2,12 @@
 <?php
 
 require realpath(__DIR__ . '/Import.php');
-require realpath(__DIR__ . '/Source/Commands/Add.php');
-require realpath(__DIR__ . '/Source/Commands/Alias.php');
-require realpath(__DIR__ . '/Source/Commands/Build.php');
-require realpath(__DIR__ . '/Source/Commands/Credential.php');
-require realpath(__DIR__ . '/Source/Commands/Flush.php');
-require realpath(__DIR__ . '/Source/Commands/Help.php');
-require realpath(__DIR__ . '/Source/Commands/Init.php');
-require realpath(__DIR__ . '/Source/Commands/Install.php');
-require realpath(__DIR__ . '/Source/Commands/Man.php');
-require realpath(__DIR__ . '/Source/Commands/Migrate.php');
-require realpath(__DIR__ . '/Source/Commands/Remove.php');
-require realpath(__DIR__ . '/Source/Commands/Run.php');
-require realpath(__DIR__ . '/Source/Commands/Serve.php');
-require realpath(__DIR__ . '/Source/Commands/Update.php');
-require realpath(__DIR__ . '/Source/Commands/Version.php');
-require realpath(__DIR__ . '/Source/Commands/Watch.php');
 
-use Phpkg\Classes\Credential\Credential;
-use Phpkg\Classes\Credential\Credentials;
 use Phpkg\Classes\Environment\Environment;
-use Phpkg\Exception\CredentialCanNotBeSetException;
 use PhpRepos\Cli\IO\Read;
-use PhpRepos\FileManager\File;
-use PhpRepos\FileManager\JsonFile;
 use PhpRepos\FileManager\Path;
 use function Phpkg\Exception\register_exception_handler;
-use function Phpkg\Providers\GitHub\github_token;
-use const Phpkg\Providers\GitHub\GITHUB_DOMAIN;
+use function PhpRepos\FileManager\File\exists;
 
 register_exception_handler();
 
@@ -41,60 +19,12 @@ if (is_null($command)) {
 
 $environment = new Environment(Path::from_string(__DIR__));
 
-match ($command) {
-    'add' => \Phpkg\Commands\Add\run($environment),
-    'alias' => \Phpkg\Commands\Alias\run($environment),
-    'build' => \Phpkg\Commands\Build\run($environment),
-    'credential' => \Phpkg\Commands\Credential\run($environment),
-    'flush' => \Phpkg\Commands\Flush\run($environment),
-    'init' => \Phpkg\Commands\Init\run($environment),
-    'install' => \Phpkg\Commands\Install\run($environment),
-    'man' => \Phpkg\Commands\Man\run(),
-    'migrate' => \Phpkg\Commands\Migrate\run($environment),
-    'remove' => \phpkg\Commands\Remove\run($environment),
-    'run' => \phpkg\Commands\Run\run($environment),
-    'serve' => \phpkg\Commands\Serve\run($environment),
-    'update' => \Phpkg\Commands\Update\run($environment),
-    'version' => \Phpkg\Commands\Version\run(),
-    'watch' => \Phpkg\Commands\Watch\run(),
-    default => \Phpkg\Commands\Help\run(),
-};
+$command = ucfirst($command);
 
-function set_credentials(Environment $environment): void
-{
-    $environment_token = github_token();
-
-    if (strlen($environment_token) > 0) {
-        return;
-    }
-
-    if (! File\exists($environment->credential_file)) {
-        throw new CredentialCanNotBeSetException('There is no credential file. Please use the `credential` command to add your token.');
-    }
-
-    /** @var Credential $github_credential */
-    $github_credential = Credentials::from_array(JsonFile\to_array($environment->credential_file))
-        ->first(fn (Credential $credential) => $credential->provider() === GITHUB_DOMAIN);
-
-    github_token(is_null($github_credential) ? '' : $github_credential->token());
+if (exists(__DIR__ . "/Source/Commands/$command.php")) {
+    $command = require realpath(__DIR__ . "/Source/Commands/$command.php");
+} else {
+    $command = require realpath(__DIR__ . '/Source/Commands/Help.php');
 }
 
-function when(bool $condition, Closure $then, Closure $otherwise = null): mixed
-{
-    return $condition ? $then() : (is_null($otherwise) ? null : $otherwise());
-}
-
-function unless(bool $condition, Closure $then, Closure $otherwise = null): mixed
-{
-    return when(! $condition, $then, $otherwise);
-}
-
-function pipe(mixed $value, Closure $closure): mixed
-{
-    return is_callable($value) ? $closure($value()) : $closure($value);
-}
-
-function when_exists(mixed $value, Closure $then, Closure $otherwise = null): mixed
-{
-    return pipe($value, is_null($value) ? (is_null($otherwise) ? fn () => null : $otherwise) : $then);
-}
+$command($environment);
